# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/addingFunds": {
            "post": {
                "description": "Add funds user by id",
                "tags": [
                    "User"
                ],
                "summary": "AddingFunds",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/createOrder": {
            "post": {
                "description": "Create order by user_id, service_id, amount",
                "tags": [
                    "User"
                ],
                "summary": "CreateOrder",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/getBalance": {
            "get": {
                "description": "Get user balance by id",
                "tags": [
                    "User"
                ],
                "summary": "GetBalance",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/getTransactions": {
            "get": {
                "description": "get transactions by user id",
                "tags": [
                    "User"
                ],
                "summary": "GetTransactions",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/revenueRecognition": {
            "delete": {
                "description": "Delete user from db",
                "tags": [
                    "User"
                ],
                "summary": "DeleteUser",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "patch": {
                "description": "Debiting funds from a temporary account to the company's bank",
                "tags": [
                    "User"
                ],
                "summary": "RevenueRecognition",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    }
}